+++ src/adapters/stargate/tests/testCases.ts
@@ -1,6 +1,6 @@
 import { Chain } from '../../../core/constants/chains'
 import { TimePeriod } from '../../../core/constants/timePeriod'
-import type { TestCase } from '../../testCase'
+import type { TestCase } from '../../../types/testCase'
 
 export const testCases: TestCase[] = [
   {
